requirements.txt
Django==4.1.10
django-environ==0.9.0
psycopg2==2.9.5
pdfminer.six==20221105
